% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designBased.r
\name{designBased}
\alias{designBased}
\title{Compute Design Based Draper-Gittoes Z-Scores}
\usage{
designBased(gamma, outcome_vec, pcf_cat_vec, instid_vec)
}
\arguments{
\item{gamma}{TODO}

\item{outcome_vec}{TODO}

\item{pcf_cat_vec}{TODO}

\item{instid_vec}{TODO}
}
\description{
\code{designBased} implements design based approach and
computes O, E, and z vectors following Draper-Gittoes (2004)
}
