% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitBabyMonitor.R
\name{fitBabyMonitor}
\alias{fitBabyMonitor}
\title{Fit Baby-MONITOR for CPQCC/VON}
\usage{
fitBabyMonitor(minimal_data, num_cat, num_cont, var_intercept, var_inst,
  var_cat, var_cat_2way, var_cont, var_subset = 1, var_inst_subset_2way = 1,
  subset = FALSE, fit_method = "probit", sparse = TRUE, burn_in = 100,
  iters = 1000, alpha = 0.1, verbose = TRUE)
}
\description{
\code{fitBabyMonitor} comprehensivly applys the Baby-MONITOR score
}
