% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullSimulate.R
\name{nullSimulate}
\alias{nullSimulate}
\title{Null Simulations for D-G Z-Scores}
\usage{
nullSimulate(minimal_data, num_cat, num_cont, var_intercept, var_inst, var_cat,
  var_cat_2way, var_cont, var_subset = 1, var_inst_subset_2way = 1,
  visualize = TRUE, mc_iters = 10, iters = 10, burn_in = 10)
}
\arguments{
\item{minimal_data}{TODO}

\item{num_cat}{TODO}

\item{num_cont}{TODO}

\item{var_intercept}{TODO}

\item{var_inst}{TODO}

\item{var_cat}{TODO}

\item{var_cat_2way}{TODO}

\item{var_cont}{TODO}

\item{var_subset}{TODO}

\item{var_inst_subset_2way}{TODO}

\item{visualize}{TODO}

\item{mc_iters}{TODO}

\item{iters}{TODO}

\item{burn_in}{TODO}
}
\description{
\code{nullSimulate} creates D-G Z-scores for a null world where all
institutions have the same latent quality measure.
}
