% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions.R
\name{pointQuantile}
\alias{pointQuantile}
\title{Generate Point Estimates and Posterior Credible Intervals}
\usage{
pointQuantile(x_mat, alpha)
}
\arguments{
\item{x_mat}{<Ixp> matrix of MCMC iterations}

\item{alpha}{real valued scalar in [0,1]}
}
\description{
\code{pointQuantile} Returns a list with the point estimate
and the upper and lower bounds for a 1-alpha "\%" credible interval.
}
