% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions.R
\name{posteriorCredible}
\alias{posteriorCredible}
\title{Compute Credible Interval and Point Estimate}
\usage{
posteriorCredible(mcmc_mat, alpha = 0.05, est_fun = median)
}
\arguments{
\item{mcmc_mat}{TODO}

\item{alpha}{TODO}

\item{est_fun}{TODO}
}
\description{
TODO There is redundancy between this and pointQuantile. resolve into 1 function.
}
