% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probitFit.r
\name{probitFit}
\alias{probitFit}
\title{MCMC Samples of Bayesian Probit Regression}
\usage{
probitFit(y, X, prior_var_vec, iters = 100, diagnostic = TRUE)
}
\description{
\code{probitFit} generates  MCMC samples of Probit regression, 
following Albert and Chib (1993)
Uses truncnorm package to sample from truncated normal	
Assumes prior centered at 0, and diagonal prior covariance structure
Assumes response y is coded as 0-1
}
