% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toDG.R
\name{toDG}
\alias{toDG}
\title{Convert MCMC Coefficients to Z-Scores Following Draper-Gittoes}
\usage{
toDG(coef_mat, count_vec, baseline = FALSE)
}
\arguments{
\item{coef_mat}{<IXp> coefficient matrix, with I MCMCM iterations.
The first column should be the overall intercept.}

\item{count_vec}{<px1> vector of integers, total number of observations
per institution or group.}

\item{baseline}{Logical; If TRUE, the first coeffient is constrained to be
0, and the D-G constraint is not satisfied.}
}
\description{
\code{toDG} reparamatrizes a matrix of MCMC coefficients into Draper-Gittoes
Z-Scores with a linear transformation. The dot product of the Z-scores and
the count_vec is constrained to be 0.
}
