% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianRegression.R
\name{bayesianRegression}
\alias{bayesianRegression}
\title{MCMC Samples of Bayesian Linear Regression}
\usage{
bayesianRegression(y, X, prior_var_vec, iters = 100)
}
\arguments{
\item{y}{<nX1> observed data vector}

\item{X}{<nXp> design matrix (takes objects of class matrix or sparce Matrix)}

\item{prior_var_vec}{<px1> prior variance vector}

\item{iters}{Integer; number of desired output iterations}
}
\description{
\code{bayesianRegression} generates MCMC samples from Bayesian Linear Regression.
Assumes prior mean centered at 0.
Written by Daniel Kirsner
}
