% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designBased.r
\name{designBased}
\alias{designBased}
\title{Compute Design Based Draper-Gittoes Z-Scores}
\usage{
designBased(gamma, outcome_vec, pcf_cat_vec, instid_vec)
}
\arguments{
\item{gamma}{Parameter in [0,1] controling pooling.
Gamma = 1 is complete global pooling, gamma = 0 is complete local pooling.}

\item{outcome_vec}{Vector with elements in {0,1} denoting the binary outcome.}

\item{pcf_cat_vec}{Vector with unique values for each PCF category.}

\item{instid_vec}{Vector with unique values for each institution.}
}
\description{
\code{designBased} implements design based approach and
computes O, E, and z vectors following Draper-Gittoes (2004)
}
