% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probitFit.r
\name{probitFit}
\alias{probitFit}
\title{MCMC Samples for Probit Regression}
\usage{
probitFit(y, X, prior_var_vec, iters = 100)
}
\arguments{
\item{y}{<nX1> observed data vector}

\item{X}{<nXp> design matrix (takes objects of class matrix or sparce Matrix)}

\item{prior_var_vec}{<px1> prior variance vector}

\item{iters}{Integer; number of desired output iterations}
}
\description{
\code{probitFit} generates MCMCM samples of Probit regression, following
the data-augmentation Gibbs sampler of Albert and Chib (1993).
Rquires truncnorm package to sample from truncated normal.
Assumes prior centered at 0, and diagonal prior covariance structure.
Assumes response y is coded as 0-1.
}
